services/elastic-sink/requirements.txt
kafka-python
elasticsearch
elasticsearch_dsl
